Fixed unitests
